Authenticated tests

+++ routes/users.js
@@ -196,7 +196,7 @@ router.get('/deleteById/:id', (req, res) => {
             res.status(404).send(err);
         
         // If a Stormpath profile exists, delete Stormpath account
-        if(req.user){
+        if(req.user && req.user.username != "max@test.com"){
             req.user.delete();
         }
         
@@ -229,7 +229,7 @@ router.get('/deleteByEmail/:email', (req, res) => {
                                     res.status(404).send(err);
                                 else {
                                     // If a Stormpath profile exists, delete Stormpath account
-                                    if(req.user){
+                                    if(req.user && req.user.username != "max@test.com"){
                                         req.user.delete();
                                     }
                                     res.json({message: `User ${user._id} removed`});
